--- api/swagger/swagger.go
@@ -1,3 +1,0 @@
-package swagger
-
-//go:generate go-bindata -pkg swagger -o swagger_gen.go files/
