+++ router/middleware/session/user.go
@@ -44,6 +44,10 @@ func SetUser() gin.HandlerFunc {
 			return user.Hash, err
 		})
 		if err == nil {
+			confv := c.MustGet("config")
+			if conf, ok := confv.(*model.Config); ok {
+				user.Admin = conf.IsAdmin(user)
+			}
 			c.Set("user", user)
 
 			// if this is a session token (ie not the API token)
