+++ server/user.go
@@ -1,4 +1,4 @@
-package api
+package server
 
 import (
 	"net/http"
@@ -6,31 +6,16 @@ import (
 	"github.com/gin-gonic/gin"
 
 	"github.com/drone/drone/cache"
-	"github.com/drone/drone/model"
 	"github.com/drone/drone/router/middleware/session"
 	"github.com/drone/drone/shared/crypto"
 	"github.com/drone/drone/shared/token"
 	"github.com/drone/drone/store"
 )
 
-// swagger:route GET /user user getUser
-//
-// Get the currently authenticated user.
-//
-//     Responses:
-//       200: user
-//
 func GetSelf(c *gin.Context) {
 	c.JSON(200, session.User(c))
 }
 
-// swagger:route GET /user/feed user getUserFeed
-//
-// Get the currently authenticated user's build feed.
-//
-//     Responses:
-//       200: feed
-//
 func GetFeed(c *gin.Context) {
 	repos, err := cache.GetRepos(c, session.User(c))
 	if err != nil {
@@ -46,13 +31,6 @@ func GetFeed(c *gin.Context) {
 	c.JSON(200, feed)
 }
 
-// swagger:route GET /user/repos user getUserRepos
-//
-// Get the currently authenticated user's active repository list.
-//
-//     Responses:
-//       200: repos
-//
 func GetRepos(c *gin.Context) {
 	repos, err := cache.GetRepos(c, session.User(c))
 	if err != nil {
@@ -104,28 +82,4 @@ func DeleteToken(c *gin.Context) {
 		return
 	}
 	c.String(http.StatusOK, tokenstr)
-}
-
-// swagger:response user
-type userResp struct {
-	// in: body
-	Body model.User
-}
-
-// swagger:response users
-type usersResp struct {
-	// in: body
-	Body []model.User
-}
-
-// swagger:response feed
-type feedResp struct {
-	// in: body
-	Body []model.Feed
-}
-
-// swagger:response repos
-type reposResp struct {
-	// in: body
-	Body []model.Repo
 }
